# services/api-gateway/handlers/sensortype.go
package handlers

import (
	"context"
	"encoding/json"
	"net/http"
	"strconv"
	"time"

	"github.com/go-chi/chi/v5"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"

	pb "github.com/skni-kod/iot-monitor-backend/internal/proto/sensor_service"
	"github.com/skni-kod/iot-monitor-backend/internal/types"
)

type SensorTypeHandler struct {
	client pb.SensorServiceClient
}

func NewSensorTypeHandler(client pb.SensorServiceClient) *SensorTypeHandler {
	return &SensorTypeHandler{client: client}
}

// @Summary List Sensor Types
// @Description Get a list of all sensor types
// @Tags SensorTypes
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {array} types.SensorTypeResponse "List of sensor types"
// @Failure 401 {string} string "Unauthorized"
// @Failure 500 {object} map[string]string
// @Router /api/sensor-types [get]
func (h *SensorTypeHandler) ListSensorTypes(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
	defer cancel()

	res, err := h.client.ListSensorTypes(ctx, &pb.ListSensorTypesRequest{})
	if err != nil {
		http.Error(w, "Failed to list sensor types: "+err.Error(), http.StatusInternalServerError)
		return
	}

	sensorTypeResponses := make([]types.SensorTypeResponse, 0, len(res.SensorTypes))
	for _, st := range res.SensorTypes {
		sensorTypeResponses = append(sensorTypeResponses, types.SensorTypeResponse{
			ID:           st.Id,
			Name:         st.Name,
			Model:        st.Model,
			Manufacturer: st.Manufacturer,
			Description:  st.Description,
			Unit:         st.Unit,
			MinValue:     st.MinValue,
			MaxValue:     st.MaxValue,
			CreatedAt:    st.CreatedAt.AsTime(),
		})
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(sensorTypeResponses); err != nil {
		http.Error(w, "Failed to encode response: "+err.Error(), http.StatusInternalServerError)
	}
}

// @Summary Get Sensor Type
// @Description Get details of a specific sensor type by ID
// @Tags SensorTypes
// @Produce json
// @Param id path int true "Sensor Type ID"
// @Security ApiKeyAuth
// @Success 200 {object} types.SensorTypeResponse "Sensor type details"
// @Failure 400 {object} map[string]string
// @Failure 401 {string} string "Unauthorized"
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/sensor-types/{id} [get]
func (h *SensorTypeHandler) GetSensorType(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
	defer cancel()

	idStr := chi.URLParam(r, "id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		http.Error(w, "Invalid sensor type ID", http.StatusBadRequest)
		return
	}

	res, err := h.client.GetSensorType(ctx, &pb.GetSensorTypeRequest{Id: int64(id)})
	if err != nil {
		st, ok := status.FromError(err)
		if ok && st.Code() == codes.NotFound {
			http.Error(w, "Sensor type not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Failed to get sensor type: "+err.Error(), http.StatusInternalServerError)
		return
	}

	sensorType := types.SensorTypeResponse{
		ID:           res.SensorType.Id,
		Name:         res.SensorType.Name,
		Model:        res.SensorType.Model,
		Manufacturer: res.SensorType.Manufacturer,
		Description:  res.SensorType.Description,
		Unit:         res.SensorType.Unit,
		MinValue:     res.SensorType.MinValue,
		MaxValue:     res.SensorType.MaxValue,
		CreatedAt:    res.SensorType.CreatedAt.AsTime(),
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(sensorType); err != nil {
		http.Error(w, "Failed to encode response: "+err.Error(), http.StatusInternalServerError)
	}
}

// @Summary Create Sensor Type
// @Description Create a new sensor type
// @Tags SensorTypes
// @Accept json
// @Produce json
// @Param sensorType body string true "Sensor Type Data"
// @Security ApiKeyAuth
// @Success 201 {object} string
// @Failure 400 {object} map[string]string
// @Failure 401 {string} string "Unauthorized"
// @Failure 500 {object} map[string]string
// @Router /api/sensor-types [post]
func (h *SensorTypeHandler) CreateSensorType(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
	defer cancel()

	var req pb.CreateSensorTypeRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body: "+err.Error(), http.StatusBadRequest)
		return
	}

	if req.Name == "" || req.Model == "" {
		http.Error(w, "Name and Model are required", http.StatusBadRequest)
		return
	}

	res, err := h.client.CreateSensorType(ctx, &req)
	if err != nil {
		http.Error(w, "Failed to create sensor type: "+err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	if err := json.NewEncoder(w).Encode(res.SensorType); err != nil {
		http.Error(w, "Failed to encode response: "+err.Error(), http.StatusInternalServerError)
	}
}

// @Summary Update Sensor Type
// @Description Update an existing sensor type by ID
// @Tags SensorTypes
// @Accept json
// @Produce json
// @Param id path int true "Sensor Type ID"
// @Param sensorType body string true "Sensor Type Data"
// @Security ApiKeyAuth
// @Success 200 {object} string
// @Failure 400 {object} map[string]string
// @Failure 401 {string} string "Unauthorized"
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/sensor-types/{id} [put]
func (h *SensorTypeHandler) UpdateSensorType(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
	defer cancel()

	idStr := chi.URLParam(r, "id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		http.Error(w, "Invalid sensor type ID", http.StatusBadRequest)
		return
	}

	var req pb.UpdateSensorTypeRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body: "+err.Error(), http.StatusBadRequest)
		return
	}

	req.Id = int64(id)

	if req.Name == "" || req.Model == "" {
		http.Error(w, "Name and Model are required", http.StatusBadRequest)
		return
	}

	res, err := h.client.UpdateSensorType(ctx, &req)
	if err != nil {
		st, ok := status.FromError(err)
		if ok && st.Code() == codes.NotFound {
			http.Error(w, "Sensor type not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Failed to update sensor type: "+err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(res.SensorType); err != nil {
		http.Error(w, "Failed to encode response: "+err.Error(), http.StatusInternalServerError)
	}
}

// @Summary Delete Sensor Type
// @Description Delete a sensor type by ID
// @Tags SensorTypes
// @Param id path int true "Sensor Type ID"
// @Success 204
// @Security ApiKeyAuth
// @Failure 400 {object} map[string]string
// @Failure 401 {string} string "Unauthorized"
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/sensor-types/{id} [delete]
func (h *SensorTypeHandler) DeleteSensorType(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
	defer cancel()

	idStr := chi.URLParam(r, "id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		http.Error(w, "Invalid sensor type ID", http.StatusBadRequest)
		return
	}

	_, err = h.client.DeleteSensorType(ctx, &pb.DeleteSensorTypeRequest{Id: int64(id)})
	if err != nil {
		st, ok := status.FromError(err)
		if ok && st.Code() == codes.NotFound {
			http.Error(w, "Sensor type not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Failed to delete sensor type: "+err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}